[fix] (erc20) Return details of token transfer when instantiating token utils
