+++ packages/coinlib-ethereum/src/types.ts
@@ -67,6 +67,7 @@ export const EthereumPaymentsUtilsConfig = extendCodec(
     decimals: t.number,
     providerOptions: t.any,
     web3: t.any,
+    tokenAddress: t.string,
   },
   'EthereumPaymentsUtilsConfig'
 )
