+++ packages/coinlib-ethereum/test/erc20/end-to-end.test.ts
@@ -1,3 +1,5 @@
+import { BaseTransactionInfo } from './../../../coinlib-common/src/types';
+import { EthereumPaymentsUtils } from '@bitaccess/coinlib-ethereum';
 import BigNumber from 'bignumber.js'
 import server from 'ganache-core'
 import {
@@ -6,7 +8,7 @@ import {
   NetworkType,
 } from '@bitaccess/coinlib-common'
 
-import { TestLogger } from '../../../../common/testUtils'
+import { expectEqualOmit, TestLogger } from '../../../../common/testUtils'
 
 import { hdAccount } from '../fixtures/accounts'
 import { HdEthereumPayments, HdErc20Payments, HdErc20PaymentsConfig, EthereumPaymentsFactory } from '../../src'
@@ -47,16 +49,20 @@ const tokenDistributor = {
 
 const target = { address: '0x62b72782415394f1518da5ec4de6c4c49b7bf854'} // payport 1
 
-const TOKEN_CONFIG = {
+const TOKEN_UTILS_CONFIG = {
   network: NetworkType.Testnet,
   fullNode: `${LOCAL_NODE}:${LOCAL_PORT}`,
   parityNode: 'none',
   gasStation: 'none',
-  hdKey: tokenIssuer.xkeys.xprv,// hdAccount.root.KEYS.xprv,
   logger,
   abi: CONTRACT_JSON,
   tokenAddress: '',
   decimals: 7,
+}
+
+const TOKEN_CONFIG = {
+  ...TOKEN_UTILS_CONFIG,
+  hdKey: tokenIssuer.xkeys.xprv,// hdAccount.root.KEYS.xprv,
   depositKeyIndex: 0,
 }
 
@@ -74,6 +80,8 @@ const HD_CONFIG = {
   decimals: 7,
   depositKeyIndex: 0,
 }
+
+const PROXY_SWEEP_AMOUNT = '163331000'
 
 jest.setTimeout(100000)
 describe('end to end tests', () => {
@@ -104,6 +112,7 @@ describe('end to end tests', () => {
   })
 
   describe('HD payments', () => {
+    let tokenAddress: string
     test('deploy erc20 contract and send funds from distribution account to hd', async () => {
 
       // deploy erc20 contract
@@ -113,18 +122,18 @@ describe('end to end tests', () => {
       const deployedContract = await ethereumHD.broadcastTransaction(signedContractDeploy)
       const contractInfo = await ethereumHD.getTransactionInfo(deployedContract.id)
       const data: any = contractInfo.data
-      const contractAddress = data.contractAddress
+      tokenAddress = data.contractAddress
 
       const tokenHD = factory.newPayments({
         ...TOKEN_CONFIG,
-        tokenAddress: contractAddress,
+        tokenAddress,
       } as HdErc20PaymentsConfig)
 
       const unsignedTx = await tokenHD.createTransaction(0, { address: source.address }, '6500000000')
       const signedTx = await tokenHD.signTransaction(unsignedTx)
       const broadcastedTx = await tokenHD.broadcastTransaction(signedTx)
 
-      HD_CONFIG.tokenAddress = contractAddress
+      HD_CONFIG.tokenAddress = tokenAddress
       hd = factory.newPayments(HD_CONFIG as HdErc20PaymentsConfig)
 
       const { confirmedBalance: confirmedDistributorBalance } = await hd.getBalance(tokenDistributor.address)
@@ -179,7 +188,7 @@ describe('end to end tests', () => {
       const preBalanceSource = await hd.getBalance(source.address)
       const preConfig = hd.getFullConfig()
 
-      const unsignedTx = await hd.createTransaction(0, { address: destination }, '163331000')
+      const unsignedTx = await hd.createTransaction(0, { address: destination }, PROXY_SWEEP_AMOUNT)
       const signedTx = await hd.signTransaction(unsignedTx)
 
       const broadcastedTx = await hd.broadcastTransaction(signedTx)
@@ -188,17 +197,19 @@ describe('end to end tests', () => {
       const { confirmedBalance: balanceSource } = await hd.getBalance(source.address)
       const { confirmedBalance: balanceTarget } = await hd.getBalance(destination)
 
-      expect(balanceTarget).toEqual('163331000')
+      expect(balanceTarget).toEqual(PROXY_SWEEP_AMOUNT)
       expect(balanceSource).toEqual('6336669000')
       expect(preConfig).toEqual(hd.getFullConfig())
-      expect(txInfo.amount).toEqual('163331000')
+      expect(txInfo.amount).toEqual(PROXY_SWEEP_AMOUNT)
     })
+
+    let sweepTxInfo: BaseTransactionInfo
 
     test('sweep transaction to contract address', async () => {
       const destination = target.address
 
       const { confirmedBalance: balanceSourcePre } = await hd.getBalance(depositAddresses[0])
-      expect(balanceSourcePre).toEqual('163331000')
+      expect(balanceSourcePre).toEqual(PROXY_SWEEP_AMOUNT)
       const { confirmedBalance: balanceTargetPre } = await hd.getBalance(target)
       expect(balanceTargetPre).toEqual('0')
 
@@ -206,6 +217,10 @@ describe('end to end tests', () => {
       const signedTx = await hd.signTransaction(unsignedTx)
       const broadcastedTx = await hd.broadcastTransaction(signedTx)
       let txInfo = await hd.getTransactionInfo(broadcastedTx.id)
+      sweepTxInfo = txInfo
+      expect(txInfo.fromAddress).toBe(depositAddresses[0])
+      expect(txInfo.toAddress).toBe(destination)
+      expect(txInfo.amount).toBe(PROXY_SWEEP_AMOUNT)
 
       expect((txInfo.toAddress || '')).toBe(destination.toLowerCase())
       expect((txInfo.fromAddress || '')).toBe(depositAddresses[0].toLowerCase())
@@ -223,7 +238,7 @@ describe('end to end tests', () => {
       const { confirmedBalance: balanceTarget } = await hd.getBalance(destination)
 
       expect(balanceSource).toEqual('0')
-      expect(balanceTarget).toEqual('163331000')
+      expect(balanceTarget).toEqual(PROXY_SWEEP_AMOUNT)
     })
 
     test('sweep from hot wallet', async () => {
@@ -234,6 +249,8 @@ describe('end to end tests', () => {
 
       const broadcastedTx = await hd.broadcastTransaction(signedTx)
       const txInfo = await hd.getTransactionInfo(broadcastedTx.id)
+      expect(txInfo.toAddress).toBe(destination)
+      expect(txInfo.amount).toBe(unsignedTx.amount)
 
       const { confirmedBalance: balanceSource } = await hd.getBalance(0)
       const { confirmedBalance: balanceTarget } = await hd.getBalance(destination)
@@ -251,6 +268,21 @@ describe('end to end tests', () => {
         sweepable: false,
         requiresActivation: false,
       })
+    })
+
+    test('erc20 utils getTxInfo returns expected values', async () => {
+      const utils = factory.newUtils({
+        ...TOKEN_UTILS_CONFIG,
+        tokenAddress,
+      })
+      // Expect utils to return the same result as payments
+      const paymentsInfo = await hd.getTransactionInfo(sweepTxInfo.id)
+      const utilsInfo = await utils.getTransactionInfo(sweepTxInfo.id)
+      expectEqualOmit(utilsInfo, paymentsInfo, ['confirmations', 'currentBlockNumber', 'data.currentBlock'])
+      // Expect utils to return info for the erc20 transfer, not the base 0-value eth fields
+      expect(utilsInfo.fromAddress).toBe(depositAddresses[0])
+      expect(utilsInfo.toAddress).toBe(target.address)
+      expect(utilsInfo.amount).toBe(PROXY_SWEEP_AMOUNT)
     })
   })
 })
